Working on removing warnings related to sizes on 64 bit architectures.
  1. Replaced unsigned long with size_t for return value of
     AT_calcAllocatedBytes in memory.c
  2. Replaced int with ptrdiff_t for one variable in gc.c
       - this is only a start; needs much more work
  3. Added an assert to check for size overflow in bafio.c
       - likely more needed elsewhere in this file and others
  4. In aterm.c
       - changed unsigned int and unsigned long to size_t where related to size
       - all Clang warning messages fixed for aterm.c

+++ aterm/aterm/gc.c
@@ -728,7 +728,7 @@ void major_sweep_phase_old()
       int alive_in_block = 0;
       int dead_in_block  = 0;
       int free_in_block  = 0;
-      int capacity = ((block->end)-(block->data))/size;
+      ptrdiff_t capacity = ((block->end)-(block->data))/size; /*RASMUSSEN - can't be int on 64 bit architectures */
       header_type *cur;
 
       assert(block->size == size);
